Refactor logout method in AuthController for improved readability by directly passing the authenticated user to the logout service.

+++ app/Http/Controllers/Api/V1/UserController.php
@@ -44,7 +44,7 @@ class UserController extends Controller
      */
     public function updateProfile(UpdateProfileRequest $request): JsonResponse
     {
-        $data = $this->userService->update((int) auth()->id(), $request->validated());
+        $data = $this->userService->update(auth()->id(), $request->validated());
 
         return $this->success($data);
     }
